feat: Add previously selected with "Change" link
If previously selected is specified then the value
will be displayed with a "Change" link. This is
ideal for a multi step process when a user may
select a country before using entity search. If
they decide the country is wrong then they can
jump back to the step to change it.