feat: Add inset text for known entites
As an example:

When a company is know then inset text is
displayed so that users can browse to the company
page and add activity.

When a company is not known then we need to handle
the callback and move to the next step which will
be confirming company details.